# api.yaml
openapi: 3.0.1
info:
  title: weather_backend
  version: v1
paths:
...
  /city/all:
    get:
      tags:
        - City
      parameters:
        - name: num
          in: query
          schema:
            type: integer
            format: int32
            default: 100
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
  '/city/{name}':
    get:
      tags:
        - City
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/encryption/encrypt:
    get:
      tags:
        - Encryption
      parameters:
        - name: Message
          in: query
          required: true
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/encryption/decrypt:
    post:
      tags:
        - Encryption
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecryptMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DecryptMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DecryptMessageRequest'
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /geolocation/ipaddress:
    get:
      tags:
        - Geolocation
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /geolocation/location:
    get:
      tags:
        - Geolocation
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  '/weather/city/{cityId}':
    get:
      tags:
        - WeatherForecast
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: number
            format: double
        - name: units
          in: query
          schema:
            $ref: '#/components/schemas/WeatherUnit'
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    City:
      required:
        - coord
        - country
        - id
        - name
        - state
      type: object
      properties:
        id:
          type: number
          format: double
        name:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        coord:
          $ref: '#/components/schemas/Coordinate'
      additionalProperties: false
    Cloud:
      type: object
      properties:
        all:
          type: integer
          format: int32
      additionalProperties: false
    Coordinate:
      type: object
      properties:
        lon:
          type: number
          format: double
        lat:
          type: number
          format: double
      additionalProperties: false
...
    DecryptMessageRequest:
      required:
        - key
        - message
        - nonce
      type: object
      properties:
        message:
          type: string
          nullable: true
        key:
          type: string
          nullable: true
        nonce:
          type: string
          nullable: true
      additionalProperties: false
    MainWeather:
      type: object
      properties:
        temp:
          type: number
          format: double
        feels_like:
          type: number
          format: double
        temp_min:
          type: number
          format: double
        temp_max:
          type: number
          format: double
        pressure:
          type: integer
          format: int32
        humidity:
          type: integer
          format: int32
        sea_level:
          type: integer
          format: int32
          nullable: true
        grnd_level:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    Rain:
      type: object
      properties:
        1h:
          type: number
          format: double
          nullable: true
        3h:
          type: number
          format: double
          nullable: true
      additionalProperties: false
...
    Snow:
      type: object
      properties:
        1h:
          type: number
          format: double
          nullable: true
        3h:
          type: number
          format: double
          nullable: true
      additionalProperties: false
...
    Weather:
      required:
        - description
        - icon
        - id
        - main
      type: object
      properties:
        id:
          type: integer
          format: int32
        main:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
      additionalProperties: false
    WeatherData:
      required:
        - clouds
        - coord
        - main
        - name
        - sys
        - weather
        - wind
      type: object
      properties:
        coord:
          $ref: '#/components/schemas/Coordinate'
        weather:
          type: array
          items:
            $ref: '#/components/schemas/Weather'
          nullable: true
        base:
          type: string
          nullable: true
        main:
          $ref: '#/components/schemas/MainWeather'
        visibility:
          type: integer
          format: int32
        wind:
          $ref: '#/components/schemas/Wind'
        clouds:
          $ref: '#/components/schemas/Cloud'
        rain:
          $ref: '#/components/schemas/Rain'
        snow:
          $ref: '#/components/schemas/Snow'
        dt:
          type: integer
          format: int32
        sys:
          $ref: '#/components/schemas/WeatherSystem'
        timezone:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        cod:
          type: integer
          format: int32
      additionalProperties: false
    WeatherSystem:
      required:
        - country
        - sunrise
        - sunset
      type: object
      properties:
        type:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        country:
          type: string
          nullable: true
        sunrise:
          type: integer
          format: int32
        sunset:
          type: integer
          format: int32
      additionalProperties: false
    WeatherUnit:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
    Wind:
      type: object
      properties:
        speed:
          type: number
          format: double
        deg:
          type: integer
          format: int32
        gust:
          type: number
          format: double
          nullable: true
      additionalProperties: false